requirements.txt
pvporcupine
pyaudio
pyttsx3
SpeechRecognition
openai-whisper
wikipedia
python-dotenv